projects/workload_scoring/requirements.txt
pandas-gbq
google-auth

python-dotenv
